[이유호] sprint 4

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicChannelService.java
@@ -1,34 +1,37 @@
 package com.sprint.mission.discodeit.service.basic;
 
+import com.sprint.mission.discodeit.dto.channel.ChannelCreateDto;
 import com.sprint.mission.discodeit.entity.Channel;
-import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.repository.ChannelRepository;
+import com.sprint.mission.discodeit.repository.MessageRepository;
+import com.sprint.mission.discodeit.repository.ReadStatusRepository;
 import com.sprint.mission.discodeit.service.ChannelService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
+@Service
+@RequiredArgsConstructor
 public class BasicChannelService implements ChannelService {
 
     private final ChannelRepository channelRepository;
-
-    public BasicChannelService(ChannelRepository channelRepository) {
-        this.channelRepository = channelRepository;
-    }
-
+    private final MessageRepository messageRepository;
+    private final ReadStatusRepository readStatusRepository;
 
     @Override
-    public Channel createChannel(String name) {
-        Channel channel = new Channel(name);
+    public Channel create(ChannelCreateDto dto) {
+        // 모든 입력을 받았다고 가정
+        Channel channel = new Channel(dto.type(), dto.name(), dto.description());
         return channelRepository.save(channel);
     }
 
     @Override
     public Channel find(UUID channelId) {
         return channelRepository.findById(channelId)
-                .orElseThrow(() -> new NoSuchElementException("채널을 찾을 수 없습니다. ID : " + channelId));
+                        .orElseThrow(() -> new NoSuchElementException("Channel with id " + channelId + " not found"));
     }
 
     @Override
@@ -37,21 +40,18 @@ public class BasicChannelService implements ChannelService {
     }
 
     @Override
-    public List<Channel> findByUser(User user) {
-        return channelRepository.findAll()
-                .stream()
-                .filter(channel -> channel.getUsers().stream()
-                        .anyMatch(u -> u.getId().equals(user.getId())))
-                .collect(Collectors.toList());
-    }
-
-    @Override
-    public void update(Channel channel) {
-        channelRepository.save(channel);
+    public Channel update(UUID channelId, String newName, String newDescription) {
+        Channel channel = channelRepository.findById(channelId)
+                .orElseThrow(() -> new NoSuchElementException("Channel with id " + channelId + " not found"));
+        channel.update(newName, newDescription);
+        return channelRepository.save(channel);
     }
 
     @Override
     public void delete(UUID channelId) {
+        if (!channelRepository.existsById(channelId)) {
+            throw new NoSuchElementException("Channel with id " + channelId + " not found");
+        }
         channelRepository.deleteById(channelId);
     }
 }
